efficientdet/requirements.txt
absl-py>=0.7.1
matplotlib>=3.0.3
numpy>=1.16.4,<1.19
Pillow>=6.0.0
PyYAML>=5.1
six>=1.12.0
tensorflow>=2.1.0
tensorflow-addons>=0.9.1
tensorflow-probability>=0.9.0
